move extract host to utils and add test
